refactor: merged generateColorMap to RasterClassifyLegend
