+++ sites/geohub/src/lib/helper/index.ts
@@ -3,7 +3,6 @@ export * from './fetchUrl';
 export * from './fetchWithTimeout';
 export * from './expression2cql';
 export * from './expression2fields';
-export * from './generateColorMap';
 export * from './getAccessLevelIcon';
 export * from './getActiveBandIndex';
 export * from './getBase64EncodedUrl';
